requirements.txt
flask==3.1.1
python-dateutil==2.9.0
peewee==3.18.1
flasgger==0.9.7.1
python-dotenv==1.0.0
requests==2.32.4